requirements.txt
requests
rich
python-dotenv
ldap3